+++ src/solvers.jl
@@ -36,11 +36,10 @@ end
 Compute the nonlinear residual for the current solution.
 """
 function compute_nonlinear_residual!(residual, A, b, sol, unknowns, PD, SC, freedofs)
-    fill!(residual.entries, 0)
+    residual.entries .= b.entries
     for j in 1:length(b), k in 1:length(b)
-        addblock_matmul!(residual[j], A[j, k], sol[unknowns[k]])
+        addblock_matmul!(residual[j], A[j, k], sol[unknowns[k]], factor = -1)
     end
-    residual.entries .-= b.entries
 
     for op in PD.operators
         residual.entries[fixed_dofs(op)] .= 0
@@ -221,39 +220,140 @@ Solves the linear system and updates the solution vector. This includes:
 - Computing the residual
 - Updating the solution with optional damping
 """
-function solve_linear_system!(A, b, sol, soltemp, residual, linsolve, unknowns, freedofs, damping, PD, SC, stats, is_linear, timer)
-    # Update system matrix if needed
-    if !SC.parameters[:constant_matrix] || !SC.parameters[:initialized]
-        if length(freedofs) > 0
-            linsolve.A = A.entries.cscmatrix[freedofs, freedofs]
-        else
-            linsolve.A = A.entries.cscmatrix
+function solve_linear_system!(A, b, sol, soltemp, residual, linsolve, unknowns, freedofs, damping, PD, SC, stats, is_linear, timer, kwargs...)
+
+    @timeit timer "assembly" begin
+        ## assemble restrctions
+        if !SC.parameters[:initialized]
+            for restriction in PD.restrictions
+                @timeit timer "$(restriction.parameters[:name])" assemble!(restriction, sol, SC; kwargs...)
+            end
         end
     end
 
-    # Set right-hand side
-    if length(freedofs) > 0
-        linsolve.b = residual.entries[freedofs]
-    else
-        linsolve.b = residual.entries
-    end
-    SC.parameters[:initialized] = true
+    @timeit timer "linear solver" begin
 
-    # Solve linear system
-    push!(stats[:matrix_nnz], nnz(linsolve.A))
-    @timeit timer "solve! call" Δx = LinearSolve.solve!(linsolve)
+        # does the linsolve object need a (new) matrix?
+        linsolve_needs_matrix = !SC.parameters[:constant_matrix] || !SC.parameters[:initialized]
+
+        ## start with the assembled matrix containing all assembled operators
+        if linsolve_needs_matrix
+            if length(freedofs) > 0
+                A_unrestricted = A.entries.cscmatrix[freedofs, freedofs]
+            else
+                A_unrestricted = A.entries.cscmatrix
+            end
+        end
+
+        # we solve for A Δx = r
+        # and update x = sol - Δx
+        if length(freedofs) > 0
+            b_unrestricted = residual.entries[freedofs]
+        else
+            b_unrestricted = residual.entries
+        end
+
+        ## restrictions only involve the blocks coressponding to the unknowns and not necessarily the full sol.entries
+        sol_freedofs = mortar([view(sol[u]) for u in unknowns])
+
+        if length(PD.restrictions) == 0
+            if linsolve_needs_matrix
+                linsolve.A = A_unrestricted
+            end
+            linsolve.b = b_unrestricted
+        else
+            # add possible Lagrange restrictions
+            restriction_matrices = [length(freedofs) > 0 ? view(restriction_matrix(re), freedofs, :) : restriction_matrix(re) for re in PD.restrictions ]
+            restriction_rhss = deepcopy([length(freedofs) > 0 ? view(restriction_rhs(re), freedofs) : restriction_rhs(re) for re in PD.restrictions ])
+
+            # block sizes for the block matrix
+            block_sizes = [size(A_unrestricted, 2), [ size(B, 2) for B in restriction_matrices ]...]
+
+            # total number of additional LM dofs
+            nLMs = @views sum(block_sizes[2:end])
+
+            @timeit timer "LM restrictions (nLMs = $nLMs)" begin
+
+                ## we need to add the (initial) solution to the rhs, since we work with the residual equation
+                for (B, rhs) in zip(restriction_matrices, restriction_rhss)
+                    rhs .-= B'sol_freedofs
+                end
+
+                total_size = sum(block_sizes)
+                Tv = eltype(A_unrestricted)
+
+                ## create block matrix
+                if linsolve_needs_matrix
+                    A_block = BlockMatrix(spzeros(Tv, total_size, total_size), block_sizes, block_sizes)
+                    A_block[Block(1, 1)] = A_unrestricted
+                end
+
+                b_block = BlockVector(zeros(Tv, total_size), block_sizes)
+
+                b_block[Block(1)] = b_unrestricted
+                u_unrestricted = @views linsolve.u[1:block_sizes[1]]
+                u_block = BlockVector(zeros(Tv, total_size), block_sizes)
+                u_block[Block(1)] = u_unrestricted
+
+                for i in eachindex(PD.restrictions)
+                    if linsolve_needs_matrix
+                        A_block[Block(1, i + 1)] = restriction_matrices[i]
+                        A_block[Block(i + 1, 1)] = transpose(restriction_matrices[i])
+                    end
+                    b_block[Block(i + 1)] = restriction_rhss[i]
+
+                end
+
+                linsolve.b = Vector(b_block) # convert to dense vector
+                linsolve.u = Vector(u_block) # convert to dense vector
+
+                if linsolve_needs_matrix
+
+                    # linsolve.A = sparse(A_block) # convert to CSC Matrix is very slow https://github.com/JuliaArrays/BlockArrays.jl/issues/78
+                    # do it manually:
+                    A_flat = spzeros(size(A_block))
+
+                    lasts_row = [0, axes(A_block)[1].lasts...] # add leading zero
+                    lasts_col = [0, axes(A_block)[2].lasts...] # add leading zero
+
+                    (n_row, n_col) = size(blocks(A_block))
+
+                    for i in 1:n_row, j in 1:n_col
+                        range_row = (lasts_row[i] + 1):lasts_row[i + 1]
+                        range_col = (lasts_col[j] + 1):lasts_col[j + 1]
+
+                        # write each block directly in the resulting matrix
+                        A_flat[range_row, range_col] = A_block[Block(i, j)]
+                    end
+                    linsolve.A = A_flat
+                end
+            end
+        end
+
+        SC.parameters[:initialized] = true
+
+        # Solve linear system
+        push!(stats[:matrix_nnz], nnz(linsolve.A))
+        @timeit timer "solve! call" begin
+            blocked_result = LinearSolve.solve!(linsolve)
+            blocked_Δx = blocked_result.u
+        end
+
+        # extract the solution / dismiss the lagrange multipliers
+        @views Δx = blocked_Δx[1:length(b_unrestricted)]
+    end
 
     # Compute solution update
     @timeit timer "update solution" begin
         if length(freedofs) > 0
-            x = sol.entries[freedofs] - Δx.u
+            x = sol.entries[freedofs] + Δx
         else
             x = zero(Δx)
             offset = 0
             for u in unknowns
                 ndofs_u = length(view(sol[u]))
                 x_range = (offset + 1):(offset + ndofs_u)
-                x[x_range] .= view(sol[u]) .- view(Δx, x_range)
+                x[x_range] .= view(sol[u]) .+ view(Δx, x_range)
                 offset += ndofs_u
             end
         end
@@ -261,15 +361,22 @@ function solve_linear_system!(A, b, sol, soltemp, residual, linsolve, unknowns, 
 
     # Check linear residual
     @timeit timer "linear residual computation" begin
+        residual.entries .= b.entries
         if length(freedofs) > 0
             soltemp.entries[freedofs] .= x
-            residual.entries .= A.entries.cscmatrix * soltemp.entries
+            residual.entries .-= A.entries.cscmatrix * soltemp.entries
         else
-            residual.entries .= A.entries.cscmatrix * x
+            residual.entries .-= A.entries.cscmatrix * x
         end
-        residual.entries .-= b.entries
         for op in PD.operators
             for dof in fixed_dofs(op)
+                if dof <= length(residual.entries)
+                    residual.entries[dof] = 0
+                end
+            end
+        end
+        for rs in PD.restrictions
+            for dof in fixed_dofs(rs)
                 if dof <= length(residual.entries)
                     residual.entries[dof] = 0
                 end
@@ -335,15 +442,16 @@ function solve_coupled_system!(A, b, sol, residual, linsolve, unknowns, damping,
     SC.parameters[:initialized] = true
 
     ## solve
-    Δx = LinearSolve.solve!(linsolve)
+    result = LinearSolve.solve!(linsolve)
+    Δx = result.u
 
-    # x = sol.entries - Δx.u ... in the entry ranges of the present unknowns
-    x = zero(Δx.u)
+    # x = sol.entries - Δx ... in the entry ranges of the present unknowns
+    x = zero(Δx)
     offset = 0
     for u in unknowns
         ndofs_u = length(view(sol[u]))
         x_range = (offset + 1):(offset + ndofs_u)
-        x[x_range] .= view(sol[u]) .- view(Δx.u, x_range)
+        x[x_range] .= view(sol[u]) .- view(Δx, x_range)
         offset += ndofs_u
     end
 
@@ -570,12 +678,12 @@ function CommonSolve.solve(PD::ProblemDescription, FES::Union{<:FESpace, Vector{
 
         linsolve = SC.linsolver
 
-        @timeit timer "linear solver" begin
-            linres = solve_linear_system!(
-                A, b, sol, soltemp, residual, linsolve, unknowns,
-                freedofs, damping, PD, SC, stats, is_linear, timer
-            )
-        end
+
+        linres = solve_linear_system!(
+            A, b, sol, soltemp, residual, linsolve, unknowns,
+            freedofs, damping, PD, SC, stats, is_linear, timer
+        )
+
         if SC.parameters[:verbosity] > -1
             if is_linear
                 @printf "%.3e\t" linres